refactor: 폴더 추가 및 파일이동
