+++ src/pages/_app.tsx
@@ -1,11 +1,11 @@
 import Header from '@/components/header/Header';
 import Modal from '@/components/modal/Modal';
 import Toast from '@/components/toast/Toast';
-import HalfPage from '@/containers/group/groupId/tasklist/HalfPage';
+import HalfPage from '@/containers/group/groupId/tasklist/HalfPage/HalfPage';
 import ReactQueryProviders from '@/hooks/useReactQuery';
 import '@/styles/globals.css';
 import type { AppProps } from 'next/app';
-import '../components/calendar/calendar.css';
+import '@/components/calendar/calendar.css';
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
